Make keypair cbox-aware as well...
